+++ Src/CSharpier.VSCode/CHANGELOG.md
@@ -1,3 +1,9 @@
+## [1.7.2]
+- Fix issue with csharpier server not working when localhost resolved to IPv6 ::1
+
+## [1.7.1]
+- Fix issue with csharpier server not supporting dotnet root
+
 ## [1.7.0]
 - Use CSharpier Http Server for 0.28.0+
 - Log version of CSharpier used to format a given file